relayer/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
web3==6.9.0
httpx==0.24.1
python-dotenv==1.0.0
requests==2.31.0 